+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+## [upcoming]
+- `SftpFile.write` now returns a `SftpFileWriter` that can be used to control
+  the writing process.
+
 ## 2.7.3
 - Update README.md
 - Move cli into separate package.
+- Properly handle chunk read error during stream read.
 
 ## 2.7.2+3
 - Update README.md